Activities2

+++ resources/views/user_workshops/table_user.blade.php
@@ -8,10 +8,10 @@
         <th>Fecha Inicio</th>
         <th>Fecha Fin</th>
         <th>Precio</th>
-        <th>Categorie Id</th>
-        <th>state</th>
         
-            <th colspan="3">Acciones</th>
+        <th>estado</th>
+        
+        @can('activities.destroy')<th colspan="3">Acciones</th>@endcan
         </tr>
         </thead>
         <tbody>
@@ -23,10 +23,10 @@
             <td>{{ $workshop->start }}</td>
             <td>{{ $workshop->end }}</td>
             <td>{{ $workshop->price }}</td>
-            <td>{{ $workshop->name_categorie }}</td>
+            
             <td>{{ $workshop->state }}</td>
             
-                <td width="120">
+            @can('activities.destroy')<td width="120">
                     {!! Form::open(['route' => ['workshops.destroy', $workshop->id], 'method' => 'delete']) !!}
                     <div class='btn-group'>
                         <a href="{{ route('workshops.show', [$workshop->id]) }}"
@@ -40,7 +40,7 @@
                         {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('¿estas seguro?')"]) !!}
                     </div>
                     {!! Form::close() !!}
-                </td>
+                </td>@endcan
             </tr>
         @endforeach
         </tbody>
